wip: extract angle stuff into own file
