+++ src/entity.rs
@@ -1,3 +1,4 @@
+pub mod angles;
 pub mod collision;
 pub mod collision_geometry;
 pub mod vector;
